[python] Consolidate IREP2 expr-builder helpers into shared module
The Python frontend had ~13 small IREP2 expression-builder helpers
(build_symbol, build_typecast, build_member, build_index, build_less_than,
build_greater_than, build_call_expr, build_call, build_deref_member,
build_dereference, build_address_of, contains_dyn_array) copy-pasted into
anonymous namespaces across 16 translation units. Collapse all copies into a
single source of truth, python_expr_builder.{h,cpp} (namespace python_expr),
and replace each per-TU copy with an include plus `using namespace`.

The copies came in a guarded family (dyn-array fallback + #cpp_type restore)
and a bare family. The shared version uses the guarded form everywhere -- a
strict superset: it only diverges from a bare copy when a dyn-sized array is
present (where the bare version would throw downstream) or when migrate_type
would drop #cpp_type (where the restore is strictly more faithful). Net -906
lines across the 16 files.

Behavior preservation was verified by re-running every Python regression test
under bitwuzla (the local build lacks z3/boolector): all 68 locally-checkable
previously-"failing" cases produce their expected verdict; the remainder fail
only for missing solvers (z3/boolector/--ir) or unrelated environment issues.
