+++ src/python-frontend/function_call/builtins.cpp
@@ -15,6 +15,7 @@
 #include <python-frontend/tuple_handler.h>
 #include <python-frontend/type_handler.h>
 #include <python-frontend/type_utils.h>
+#include <python-frontend/python_expr_builder.h>
 #include <util/arith_tools.h>
 #include <util/base_type.h>
 #include <util/expr_util.h>
@@ -33,105 +34,7 @@
 #include <stdexcept>
 
 using namespace json_utils;
-
-namespace
-{
-// V.3: IREP2 expression-construction helpers (exact round-trip; behaviour-
-// preserving). Back-migrated for the legacy adjust/goto-convert seam.
-//
-// Two migrate_type round-trip hazards are guarded: a dynamically-sized array
-// type (non-constant size) throws get_width downstream, and type attributes
-// such as #cpp_type are dropped. So the helpers fall back to the legacy node
-// for dyn-sized arrays / non-matching sources, and restore the exact result
-// type (result.type() = t) on the member/index/typecast nodes.
-bool contains_dyn_array(const typet &t)
-{
-  if (t.is_array())
-  {
-    const array_typet &at = to_array_type(t);
-    if (at.size().is_nil() || !at.size().is_constant())
-      return true;
-    return contains_dyn_array(at.subtype());
-  }
-  if (t.is_pointer())
-    return contains_dyn_array(t.subtype());
-  return false;
-}
-
-exprt build_symbol(const symbolt &sym)
-{
-  if (contains_dyn_array(sym.get_type()))
-    return symbol_expr(sym);
-  return migrate_expr_back(symbol_expr2tc(sym));
-}
-
-exprt build_typecast(const exprt &from, const typet &t)
-{
-  if (contains_dyn_array(t) || contains_dyn_array(from.type()))
-    return typecast_exprt(from, t);
-  expr2tc from2;
-  migrate_expr(from, from2);
-  exprt result = migrate_expr_back(typecast2tc(migrate_type(t), from2));
-  result.type() = t;
-  return result;
-}
-
-exprt build_address_of(const exprt &obj)
-{
-  if (contains_dyn_array(obj.type()))
-    return address_of_exprt(obj);
-  expr2tc obj2;
-  migrate_expr(obj, obj2);
-  return migrate_expr_back(address_of2tc(obj2->type, obj2));
-}
-
-// member2t needs a struct/union/symbol source; fall back to legacy otherwise
-// (and for dyn-array result types).
-exprt build_member(const exprt &base, const irep_idt &name, const typet &t)
-{
-  if (contains_dyn_array(t))
-    return member_exprt(base, name, t);
-  expr2tc base2;
-  migrate_expr(base, base2);
-  if (
-    is_struct_type(base2->type) || is_union_type(base2->type) ||
-    is_symbol_type(base2->type))
-  {
-    exprt result = migrate_expr_back(member2tc(migrate_type(t), base2, name));
-    result.type() = t;
-    return result;
-  }
-  return member_exprt(base, name, t);
-}
-
-// index2t needs an array/vector/symbol source; fall back to legacy otherwise
-// (and for dyn-array source/result types -- string indexing relies on the
-// #cpp_type attribute that migrate_type drops, hence result.type() = t).
-exprt build_index(const exprt &arr, const exprt &idx, const typet &t)
-{
-  if (contains_dyn_array(arr.type()) || contains_dyn_array(t))
-    return index_exprt(arr, idx, t);
-  expr2tc arr2, idx2;
-  migrate_expr(arr, arr2);
-  migrate_expr(idx, idx2);
-  if (
-    is_array_type(arr2->type) || is_vector_type(arr2->type) ||
-    is_symbol_type(arr2->type))
-  {
-    exprt result = migrate_expr_back(index2tc(migrate_type(t), arr2, idx2));
-    result.type() = t;
-    return result;
-  }
-  return index_exprt(arr, idx, t);
-}
-
-// 2-arg form: element type is the source array's subtype (matches the legacy
-// index_exprt(arr, idx) constructor).
-exprt build_index(const exprt &arr, const exprt &idx)
-{
-  return build_index(arr, idx, arr.type().subtype());
-}
-} // namespace
+using namespace python_expr;
 
 namespace
 {
